Enhance UUID base and make it base of all nodes
