+++ src/cript/nodes/subobjects/condition.py
@@ -2,11 +2,11 @@ from dataclasses import dataclass, replace
 from numbers import Number
 from typing import Union
 
-from cript.nodes.core import BaseNode
 from cript.nodes.primary_nodes.data import Data
+from cript.nodes.uuid_base import UUIDBaseNode
 
 
-class Condition(BaseNode):
+class Condition(UUIDBaseNode):
     """
     ## Definition
 
@@ -74,7 +74,7 @@ class Condition(BaseNode):
     """
 
     @dataclass(frozen=True)
-    class JsonAttributes(BaseNode.JsonAttributes):
+    class JsonAttributes(UUIDBaseNode.JsonAttributes):
         key: str = ""
         type: str = ""
         descriptor: str = ""
