+++ src/cript/nodes/subobjects/equipment.py
@@ -1,13 +1,13 @@
 from dataclasses import dataclass, field, replace
 from typing import List, Union
 
-from cript.nodes.core import BaseNode
 from cript.nodes.subobjects.citation import Citation
 from cript.nodes.subobjects.condition import Condition
 from cript.nodes.supporting_nodes.file import File
+from cript.nodes.uuid_base import UUIDBaseNode
 
 
-class Equipment(BaseNode):
+class Equipment(UUIDBaseNode):
     """
     ## Definition
     An [Equipment](https://pubs.acs.org/doi/suppl/10.1021/acscentsci.3c00011/suppl_file/oc3c00011_si_001.pdf#page=23)
@@ -42,7 +42,7 @@ class Equipment(BaseNode):
     """
 
     @dataclass(frozen=True)
-    class JsonAttributes(BaseNode.JsonAttributes):
+    class JsonAttributes(UUIDBaseNode.JsonAttributes):
         key: str = ""
         description: str = ""
         condition: List[Condition] = field(default_factory=list)
