+++ src/cript/nodes/subobjects/parameter.py
@@ -1,10 +1,10 @@
 from dataclasses import dataclass, replace
 from typing import Union
 
-from cript.nodes.core import BaseNode
+from cript.nodes.uuid_base import UUIDBaseNode
 
 
-class Parameter(BaseNode):
+class Parameter(UUIDBaseNode):
     """
     ## Definition
 
@@ -45,7 +45,7 @@ class Parameter(BaseNode):
     """
 
     @dataclass(frozen=True)
-    class JsonAttributes(BaseNode.JsonAttributes):
+    class JsonAttributes(UUIDBaseNode.JsonAttributes):
         key: str = ""
         value: Union[int, float, str] = ""
         # We explicitly allow None for unit here (instead of empty str),
