+++ src/cript/nodes/supporting_nodes/user.py
@@ -47,12 +47,10 @@ class User(UUIDBaseNode):
         all User attributes
         """
 
-        created_at: str = ""
         email: str = ""
         model_version: str = ""
         orcid: str = ""
         picture: str = ""
-        updated_at: str = ""
         username: str = ""
 
     _json_attrs: JsonAttributes = JsonAttributes()
@@ -77,10 +75,6 @@ class User(UUIDBaseNode):
         self.validate()
 
     # ------------------ properties ------------------
-
-    @property
-    def created_at(self) -> str:
-        return self._json_attrs.created_at
 
     @property
     def email(self) -> str:
@@ -121,10 +115,6 @@ class User(UUIDBaseNode):
     @property
     def picture(self) -> str:
         return self._json_attrs.picture
-
-    @property
-    def updated_at(self) -> str:
-        return self._json_attrs.updated_at
 
     @property
     def username(self) -> str:
